requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.10.0
numpy>=1.26.0
aiortc>=1.9.0
aiohttp>=3.11.0
python-dotenv>=1.0.0
